requirements.txt
absl-py==1.2.0
aiogram==2.22.1
aiohttp==3.8.3
aiosignal==1.2.0
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.2
asyncpg==0.26.0
attrs==22.1.0
audioread==3.0.0
autochord==0.1.4
Babel==2.9.1
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
decorator==5.1.1
dpcpp-cpp-rt==2022.2.0
ffprobe==0.5
filelock==3.8.0
flatbuffers==22.9.24
frozenlist==1.3.1
gast==0.4.0
gdown==4.5.1
gino==1.0.1
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.49.1
h5py==3.7.0
idna==3.4
importlib-metadata==5.0.0
intel-cmplr-lib-rt==2022.2.0
intel-cmplr-lic-rt==2022.2.0
intel-opencl-rt==2022.2.0
intel-openmp==2022.2.0
joblib==1.2.0
keras==2.10.0
Keras-Preprocessing==1.1.2
lazycats==0.1.0
libclang==14.0.6
librosa==0.9.2
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.1
mkl==2022.2.0
mkl-service==2.4.0
multidict==6.0.2
numba==0.56.2
numpy==1.23.3
oauthlib==3.2.1
opt-einsum==3.3.0
packaging==21.3
Pillow==9.2.0
pooch==1.6.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydub==0.25.1
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.2
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
soundfile==0.11.0
soupsieve==2.3.2.post1
SQLAlchemy==1.3.24
tbb==2021.7.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
threadpoolctl==3.1.0
tqdm==4.64.1
typing_extensions==4.3.0
urllib3==1.26.12
vamp==1.1.0
Werkzeug==2.2.2
wrapt==1.14.1
yarl==1.8.1
zipp==3.8.1
